homework_backend/requirements.txt
Django==5.1.3
djangorestframework==3.14.0
django-cors-headers==4.3.1
